Structs fixes.
